:truck: move: 파일 이름 수정 및 주석 추가 (#34)
- css 파일 -> 컴포넌트랑 이름 동일하게 (케밥 -> 파스칼)
- toss payments 컴폰넌트 관련 주석 추가
