+++ src/components/layout/order/footer/Footer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import useOrderStore from '@/shared/store/order';
-import styles from './styles/footer.module.css';
-import StepButtonContainer from '@/components/layout/order/Footer/StepButtonContainer';
+import styles from './styles/Footer.module.css';
+import StepButtonContainer from '@/components/layout/order/footer/StepButtonContainer';
 import { SwiperRef } from 'swiper/react';
 import { PaymentWidgetInstance } from '@tosspayments/payment-widget-sdk';
