+++ src/components/order/payment/PaymentContainer.tsx
@@ -48,20 +48,12 @@ const PaymentContainer = ({ paymentWidget }: PaymentContainerProps) => {
     });
   }, [paymentWidget]);
 
-  useEffect(() => {
-    const paymentMethodsWidget = paymentMethodsWidgetRef.current;
-
-    if (paymentMethodsWidget == null) {
-      return;
-    }
-  }, [paymentWidget]);
-
   return (
     <div className={styles.container}>
       {/* 결제 UI, 이용약관 UI 영역 */}
       <div id="payment-widget" />
       <div id="agreement" />
-      {/* 결제하기 버튼 */}
+      {/*  결제 버튼은 StepButtonContainer.tsx에 */}
     </div>
   );
 };
