+++ src/pages/order/fail.tsx
@@ -1,7 +1,11 @@
 import React from 'react';
 
-const Fail = () => {
+/**
+ * 결제 실패 페이지
+ * @constructor
+ */
+const OrderFailPage = () => {
   return <div>결제 실패...!!</div>;
 };
 
-export default Fail;
+export default OrderFailPage;
