+++ src/shared/store/order.ts
@@ -5,6 +5,13 @@ import { CategoryWithMenuItem, Tables } from '@/types/supabase';
  * 일반고객의 주문과 관련된 전역 상태를 관리하는 store 입니다.
  */
 
+export enum ORDER_STEP {
+  CHOOSE_ORDER_TYPE,
+  SELECT_MENU,
+  CHECK_MENU,
+  PAYMENT,
+}
+
 interface OrderState {
   step: number;
   readonly maxStep: number;
@@ -20,8 +27,8 @@ interface OrderState {
 
 const useOrderStore = create<OrderState>()((set, get) => ({
   //  현재 주문 단계를 나타냅니다.
-  step: 0,
-  maxStep: 4,
+  step: ORDER_STEP.CHOOSE_ORDER_TYPE,
+  maxStep: ORDER_STEP.PAYMENT,
   menuData: null,
   goNextStep: () => set(state => ({ step: Math.min(state.step + 1, state.maxStep) })),
   goPrevStep: () => set(state => ({ step: Math.max(state.step - 1, 0) })),
